cpu: Wrapping kp_block in Block.
Towards the CAPE interface.

+++ ve/cpu/codegen/buffer.cpp
@@ -1,18 +1,17 @@
 #include <sstream>
-#include <string>
 #include "utils.hpp"
 #include "codegen.hpp"
 
 using namespace std;
 using namespace bohrium::core;
 
-namespace bohrium{
+namespace bohrium {
 namespace engine{
 namespace cpu{
 namespace codegen{
 
 Buffer::Buffer(void) : buffer_(NULL), id_(0) {}
-Buffer::Buffer(kp_buffer * buffer, size_t buffer_id) : buffer_(buffer), id_(buffer_id) {
+Buffer::Buffer(kp_buffer* buffer, int64_t buffer_id) : buffer_(buffer), id_(buffer_id) {
     if (NULL == buffer_) {
         throw runtime_error("Constructing a NULL Buffer, when expecting to have one");
     }
@@ -44,12 +43,12 @@ string Buffer::etype(void)
     return etype_to_ctype_text(bhtype_to_etype(meta().type));
 }
 
-kp_buffer & Buffer::meta(void)
+kp_buffer& Buffer::meta(void)
 {
     return *buffer_;
 }
 
-uint64_t Buffer::id(void)
+int64_t Buffer::id(void)
 {
     return id_;
 }
